fix(auto-reply): defer staged media cleanup for queued turns
